Atualização nas variaveis

+++ super_trunfo.c
@@ -17,24 +17,24 @@ int main (){
     char estado[50];   //Variavel para estado
     char cdgcarta[50];  //Variavél para código de carta
     char ncidade[50];  //Variavel para cidade
-    int populacao;  //Variavel para população
+    unsigned long int populacao;  //Variavel para população
     float area;     //Variavel para area²
-    long long double pib;      //Variavel para PIB
+    long double pib;      //Variavel para PIB
     int nturisticos;  // Variavel para pontos turisticos
     float densidadepop; // Variavel para densidade populacional
-    long long double rendapercapita; // Variavel para Renda per capta
+    long double rendapercapita; // Variavel para Renda per capta
 
 
     // Variaveis da carta 2.
     char estado_b[50];   //Variavel para estado
     char cdgcarta_b[50];  //Variavél para código de carta
     char ncidade_b[50];  //Variavel para cidade
-    int populacao_b;  //Variavel para população
+    unsigned long int populacao_b;  //Variavel para população
     float area_b;     //Variavel para area²
-    long long double pib_b;      //Variavel para PIB
+    long double pib_b;      //Variavel para PIB
     int nturisticos_b;  // Variavel para pontos turisticos
     float densidadepop_b; // Variavel para densidade populacional
-    long long double rendapercapita_b; // Variavel para Renda per capta
+    long double rendapercapita_b; // Variavel para Renda per capta
 
 
     //Introdução do jogo.
@@ -51,11 +51,11 @@ int main (){
     printf ("Insira o nome da cidade \n");
     scanf ("%s", &ncidade);
     printf ("Insira o número de habitantes da cidade \n");
-    scanf ("%d", &populacao);
+    scanf ("%lu", &populacao);
     printf ("Insira a área da cidade em quilômetros quadrados \n");
     scanf ("%f", &area);
     printf ("Insira o PIB (Produto Interno Bruto da cidade) \n");
-    scanf ("%.2llf", &pib);
+    scanf ("%Lf", &pib);
     printf ("Insira a quantidade de pontos turísticos na cidade \n");
     scanf ("%d", &nturisticos);
 
@@ -74,12 +74,12 @@ int main (){
     printf ("Estado: %s\n", estado);
     printf ("Código: %s\n", cdgcarta);
     printf ("Nome da cidade: %s\n", ncidade);
-    printf ("População: %d\n", populacao);
+    printf ("População: %lu\n", populacao);
     printf ("Àrea: %.2f km²\n", area);
-    printf ("PIB: %.2llf bilhões de reais\n", pib);
+    printf ("PIB: %.2Lf bilhões de reais\n", pib);
     printf ("Números de Pontos Turísticos: %d\n", nturisticos);
     printf ("Densidade Populacional: %.2f hab/km²\n", densidadepop);
-    printf ("PIB per Capita: %.2llf reais\n", rendapercapita);
+    printf ("PIB per Capita: %.2Lf reais\n", rendapercapita);
 
     // Segunda fase inserindo os codigos da carta 2.
     
@@ -91,11 +91,11 @@ int main (){
     printf ("Insira o nome da cidade \n");
     scanf ("%s", &ncidade_b);
     printf ("Insira o número de habitantes da cidade \n");
-    scanf ("%d", &populacao_b);
+    scanf ("%lu", &populacao_b);
     printf ("Insira a área da cidade em quilômetros quadrados \n");
     scanf ("%f", &area_b);
     printf ("Insira o PIB (Produto Interno Bruto da cidade) \n");
-    scanf ("%.2llf", &pib_b);
+    scanf ("%Lf", &pib_b);
     printf ("Insira a quantidade de pontos turísticos na cidade \n");
     scanf ("%d", &nturisticos_b);
 
@@ -114,12 +114,12 @@ int main (){
     printf ("Estado: %s\n", estado_b);
     printf ("Código: %s\n", cdgcarta_b);
     printf ("Nome da cidade: %s\n", ncidade_b);
-    printf ("População: %d\n", populacao_b);
+    printf ("População: %lu\n", populacao_b);
     printf ("Àrea: %.2f km²\n", area_b);
-    printf ("PIB: %.2f bilhões de reais\n", pib_b);
+    printf ("PIB: %.2Lf bilhões de reais\n", pib_b);
     printf ("Números de Pontos Turísticos: %d\n", nturisticos_b);
     printf ("Densidade Populacional: %.2f hab/km²\n", densidadepop_b);
-    printf ("PIB per Capita: %.2f reais\n", rendapercapita_b);
+    printf ("PIB per Capita: %.2Lf reais\n", rendapercapita_b);
